修复docker中国际化错误 添加日志
