+++ src/AVOne.Server/Pages/_Layout.cshtml
@@ -49,9 +49,9 @@
     </script>
     <script>
         // disable context menu
-        document.addEventListener('contextmenu', function (e) {
-            e.preventDefault();
-        });
+        //document.addEventListener('contextmenu', function (e) {
+        //    e.preventDefault();
+        //});
     </script>
 </body>
