+++ src/AVOne.Server/wwwroot/i18n/en-US.json
@@ -180,5 +180,12 @@
         }
       }
     }
+  },
+  "System": {
+    "Title": "Advanced",
+    "Log": {
+      "Title": "Log"
+    }
   }
+  
 }
