+++ src/AVOne.Server/wwwroot/i18n/supportedCultures.json
@@ -2,5 +2,3 @@
   "zh-CN",
   "en-US"
 ]
-
-
