+++ src/AVOne.Server/wwwroot/i18n/zh-CN.json
@@ -11,23 +11,23 @@
     "Next": "下一步",
     "Cancel": "取消",
     "Submit": "提交",
-    "Disabled":"已禁用",
-    "Active":"已启用",
-    "NotSupported":"不支持",
-    "Malfunctioned":"已禁用",
-    "Restart":"待重启",
-    "On":"开启",
-    "Off":"关闭",
-    "Disable":"禁用",
-    "Enable":"启用"
+    "Disabled": "已禁用",
+    "Active": "已启用",
+    "NotSupported": "不支持",
+    "Malfunctioned": "已禁用",
+    "Restart": "待重启",
+    "On": "开启",
+    "Off": "关闭",
+    "Disable": "禁用",
+    "Enable": "启用"
   },
-  "Download":{
-    "Title":"下载",
-    "Search":"搜索任务",
-    "Sort":{
-      "ResetSort":"重置排序",
-      "SortbyName":"按名称排序",
-      "SortbyCreatedDate":"按创建时间排序"
+  "Download": {
+    "Title": "下载",
+    "Search": "搜索任务",
+    "Sort": {
+      "ResetSort": "重置排序",
+      "SortbyName": "按名称排序",
+      "SortbyCreatedDate": "按创建时间排序"
     },
     "Detail": {
       "Info": "下载信息",
@@ -95,11 +95,11 @@
       "Anime": "动画",
       "US": "欧美"
     },
-      "Status": {
-        "Downloading":"下载中",
-        "Completed": "已完成",
-        "Canceled": "已取消"
-      }
+    "Status": {
+      "Downloading": "下载中",
+      "Completed": "已完成",
+      "Canceled": "已取消"
+    }
   },
   "Settings": {
     "Title": "设置",
@@ -187,6 +187,12 @@
           "DisableRepoFailed": "禁用存储库'{0}'失败"
         }
       }
+    }
+  },
+  "System": {
+    "Title": "高级",
+    "Log": {
+      "Title": "日志"
     }
   }
 }
